OsfClient

Documentation for OsfClient.

OsfClient.get_attr

OsfClient.osf_api

OsfClient.osf_download

OsfClient.osf_download

OsfClient.osf_download

OsfClient.osf_traverse

OsfClient.get_attr — Method

get_attr(data)

returns name, kind, id and folder (either name or materialized_path)

OsfClient.osf_api — Method

osf_api(id;kwargs...)
osf_api(url;kwargs...)

Make requests against the osf api.

##args id/url: osf-id, e.g. 3a7kt or url containing "https://api.osf.io/v2/[...]". In the latter case, URL is not modified, in the former case, type and field need to be specified

kwargs

field: is put at v2/*type*/id/*field* type: is put at v2/*type* - defaults to nodes parse: if other than json returns the HTTP.get, else is converted to String and parsed with JSON.json

if parse == "json" (default), received data is put to string and parsed as json

OsfClient.osf_download — Method

osf_download(t::FileTrees.File)

download a single file and return it

OsfClient.osf_download — Method

osf_download(t::FileTrees.File, save_path::String)

download a single file, write it to the save_path and return it

OsfClient.osf_download — Method

osf_download(id::String)

download a single osf-id and return it

OsfClient.osf_traverse — Method

osf_traverse(nodeid)

osftraverses all folders and copies information about the data. Recurrent function is called `osftraverse!`